% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorgrouping.R
\name{addcategorytopolls}
\alias{addcategorytopolls}
\title{Adds Category assignment to Poll Data 
This function returns a new data frame with a priorcat column to be accessed in other functions.}
\usage{
addcategorytopolls(poll_data, proploc, stateloc, election_data,
  cutoffs = c(-0.2, -0.1, -0.025, 0.025, 0.1, 0.2),
  groupnames = c("Strong Red", "Red", "Lean Red", "Competitive",
  "Lean Blue", "Blue", "Strong Blue"))
}
\arguments{
\item{poll_data}{the data to fit the prior distribution , the first column must be the state name}

\item{proploc}{the column the adjusted poll proportion is located}

\item{stateloc}{the column the state is located in}

\item{election_data}{the election data in df form. The first column must contain the name of that state.}

\item{cutoffs}{the cutoffs used to split the data into the categories}

\item{groupnames}{labels for the categories}
}
\value{
poll_data with the prior category appended at the end
}
\description{
Adds Category assignment to Poll Data 
This function returns a new data frame with a priorcat column to be accessed in other functions.
}
\examples{
require(politicaldata)
elect2008  = subset(pres_results , year == 2008)
elect2008$margin = elect2008$dem - elect2008$rep
elect2012 = subset(pres_results , year == 2012)
elect2012$margin = elect2012$dem - elect2012$rep
electdata = data.frame("state" = elect2008$state, "2008" =  elect2008$margin, "2012" = elect2012$margin)
set.seed(16)
data(polls2016)
poll1 = polls2016[sample(1:nrow(polls2016), 500) ,]
poll2 = polls2016[sample(1:nrow(polls2016), 500), ]
propnormpolls1 = propnormdf(poll1, c(2,3))
propnormpolls2 = propnormdf(poll2, c(2,3))
addcategorytopolls(propnormpolls1, 30, 18, electdata)
newcutoff = c(-.15, -.1, -0.05, 0.05, .1, .15)
addcategorytopolls(propnormpolls2, 30, 18, electdata, cutoffs = newcutoff)
}
