% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorgrouping.R
\name{addcategorytopollsmargin}
\alias{addcategorytopollsmargin}
\title{Title}
\usage{
addcategorytopollsmargin(
  poll_data,
  marginloc,
  stateloc,
  election_data,
  cutoffs = c(-0.2, -0.1, -0.025, 0.025, 0.1, 0.2),
  groupnames = c("Strong Red", "Red", "Lean Red", "Competitive", "Lean Blue", "Blue",
    "Strong Blue")
)
}
\arguments{
\item{poll_data}{}

\item{marginloc}{}

\item{stateloc}{}

\item{election_data}{}

\item{cutoffs}{}

\item{groupnames}{}
}
\value{

}
\description{
Title
}
\examples{
require(politicaldata)
elect2008  = subset(pres_results , year == 2008)
elect2008$margin = elect2008$dem - elect2008$rep
elect2012 = subset(pres_results , year == 2012)
elect2012$margin = elect2012$dem - elect2012$rep
electdata = data.frame("state" = elect2008$state, "2008" =  elect2008$margin, "2012" = elect2012$margin)
set.seed(16)
data(polls2016)
poll1 = polls2016[sample(1:nrow(polls2016), 500) ,]
poll2 = polls2016[sample(1:nrow(polls2016), 500), ]
propnormpolls1 = propnormdf(poll1, c(2,3))
propnormpolls2 = propnormdf(poll2, c(2,3))
poll2016$margin = poll2016$Trump - polls2016$Clinton
marginloc = which(colnames(polls2016) == "margin")
stateloc = which(colnames(polls2016) == "State")
addcategorytopollsmargin(polls2016, marginloc = marginloc, stateloc = stateloc, election_data = electdata)
}
