% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorcalculation.R
\name{ae_by_state}
\alias{ae_by_state}
\title{Average Error by State}
\usage{
ae_by_state(predicted, actual)
}
\arguments{
\item{predicted}{the predictions including the first column with a character vector with idenifiers to check that the rows of predicted and actual match}

\item{actual}{the real result including the first column with a character vector}
}
\value{
a data frame with the state and the average error for that state
}
\description{
This function calculates the root mean square error for each state over a single or multiple years and checks that the idenifiers match.
}
\examples{
predict1 = data.frame(state = c("TX", "GA", "OR"), dem1 = c(.45, .49, .63), dem2 = c(.55, .51, .37))
actual1 = data.frame(state = c("TX", "GA", "OR"), dem1 = c(.42, .48, .62), dem2 = c(.58, .52, .38))
ae_by_state(predict1, actual1)
predict2 = data.frame(state = c("OK", "MS", "WY"), dem1 = c(.42, .44, .3))
actual2 = data.frame(state = c("OK", "MS", "WY"), dem1 = c(.41, .38, .3))
ae_by_state(predict2, actual2)
}
