\name{bayesurvey-package}
\alias{bayesurvey-package}
\alias{bayesurvey}
\docType{package}
\title{
  Bayesian Conjugate Prior Analysis of Survey Data
}
\description{
This package implements Bayesian conjugate prior models with a focus on handling survey data with multiple categories. It starts by normalizing and transforming the data,  then fits prior distributions, and runs a conjugate prior analysis that can be iterative. While many of the functions are designed for categorical data, the conjugate prior functions work with any type of data..
}

\author{
Brittany Alexander

Maintainer: Brittany Alexander <brittany@stat.tamu.edu>
}
\references{
Poll-Based Bayesian Models to Predict United States Presidential Elections". B. Alexander, L. Ellingson, Joint Statistical Meetings Proceedings 2019.

}
\keyword{ package }

\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
