% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{fitbetavecMOM}
\alias{fitbetavecMOM}
\title{Beta Method of Moments
Calculates the method of monements estimators of a beta distribution fitted to the data}
\usage{
fitbetavecMOM(data)
}
\arguments{
\item{data}{- a vector of data}
}
\value{
The fitted a,b of a beta(a,b) distribtion
}
\description{
Beta Method of Moments
Calculates the method of monements estimators of a beta distribution fitted to the data
}
