% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullmodel.R
\name{gaussianrollingaverageiterative}
\alias{gaussianrollingaverageiterative}
\title{Rolling Average Iterative Gaussian Bayesian Model}
\usage{
gaussianrollingaverageiterative(
  poll_data,
  stateloc,
  proploc,
  candidateloc,
  npolls = 10,
  varloc = NULL,
  nloc,
  invgamma = F,
  v0 = NULL,
  a0 = NULL,
  b0 = NULL,
  election_data,
  cutoffs = c(-0.2, -0.1, -0.025, 0.025, 0.1, 0.2),
  groupnames = c("Strong Red", "Red", "Lean Red", "Competitive", "Lean Blue", "Blue",
    "Strong Blue")
)
}
\arguments{
\item{poll_data}{}

\item{stateloc}{}

\item{proploc}{}

\item{candidateloc}{}

\item{npolls}{}

\item{varloc}{}

\item{nloc}{}

\item{invgamma}{}

\item{v0}{}

\item{a0}{}

\item{b0}{}

\item{election_data}{}

\item{cutoffs}{}

\item{groupnames}{}
}
\value{

}
\description{
Rolling Average Iterative Gaussian Bayesian Model
}
\examples{
require(politicaldata)
elect2008  = subset(pres_results , year == 2008)
elect2008$margin = elect2008$dem - elect2008$rep
elect2012 = subset(pres_results , year == 2012)
elect2012$margin = elect2012$dem - elect2012$rep
electdata = data.frame("state" = elect2008$state, "2008" =  elect2008$margin, "2012" = elect2012$margin)
data(polls2016)
polls = polls2016[ complete.cases(polls2016[, c(2, 3, 12)]), ]
head(polls)
nloc = which(colnames(polls2016) == "observations")
stateloc  = which(colnames(polls) == "State")
gaussianrollingaverage(polls, stateloc = stateloc, proploc = c(2,3), 3, nloc = nloc, election_data = electdata)
}
