% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorgrouping.R
\name{getpriordistribution}
\alias{getpriordistribution}
\title{Title}
\usage{
getpriordistribution(poll_data, proploc, election_data, cutoffs = c(-0.2,
  -0.1, -0.025, 0.025, 0.1, 0.2), groupnames = c("Strong Red", "Red",
  "Lean Red", "Competitive", "Lean Blue", "Blue", "Strong Blue"),
  catweight = NULL, stateweight = F)
}
\arguments{
\item{poll_data}{the data to fit the prior distribution , the first column must be the state name}

\item{proploc}{the column the adjusted poll proportion is located}

\item{election_data}{}

\item{cutoffs}{}

\item{groupnames}{}

\item{catweight}{}

\item{stateweight}{( maybe implement later to use a weighted average)}
}
\description{
Title
}
