% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullmodel.R
\name{iterativegaussianmodelprop}
\alias{iterativegaussianmodelprop}
\title{Iterative Gaussian Conjugate Prior Model for American Election}
\usage{
iterativegaussianmodelprop(poll_data, stateloc, proploc, candidateloc,
  varloc = NULL, nloc, election_data, cutoffs = c(-0.2, -0.1, -0.025,
  0.025, 0.1, 0.2), groupnames = c("Strong Red", "Red", "Lean Red",
  "Competitive", "Lean Blue", "Blue", "Strong Blue"))
}
\arguments{
\item{poll_data}{a data frame containing all necessary datat}

\item{stateloc}{the column number that contains the states for the poll}

\item{proploc}{the set of columns that contain poll results that need to be normalized}

\item{candidateloc}{the location of the candidate of interest}

\item{varloc}{the location of the variance if it is provided}

\item{nloc}{the location of the sample sizes}

\item{election_data}{the election data used to make the categorization decisions}

\item{cutoffs}{the cutoffs to assign the prior categories}

\item{groupnames}{the names of the prior categories}
}
\value{
a dataframe with the following rows: State with the state, Posterior Mean with the posterior mean, Posterior Variance with the posterior variance, and Posterior Standard Deviation with the posterior standard deviation.
}
\description{
This function completely fits the Iterative Gaussian Proportional model in Alexander and Ellingson (2019) given raw poll data, election data, and basic parameters.
Given polling data on congressional districts or senate seats this could make similar predictions.
}
\examples{
require(politicaldata)
elect2008  = subset(pres_results , year == 2008)
elect2008$margin = elect2008$dem - elect2008$rep
elect2012 = subset(pres_results , year == 2012)
elect2012$margin = elect2012$dem - elect2012$rep
electdata = data.frame("state" = elect2008$state, "2008" =  elect2008$margin, "2012" = elect2012$margin)
polls = polls2016[ complete.cases(polls2016[, c(2, 3, 12)]), ]
head(polls)
nloc = which(colnames(polls2016) == "observations")
stateloc  = which(colnames(polls) == "State")
iterativegaussianmodelprop(polls, stateloc, c(2,3),  3, nloc = nloc, election_data = electdata)
iterativegaussianmodelprop(polls, stateloc, c(2,3), 3, nloc = nloc, cutoffs = c(-.15, -.1, -0.05, 0.05, .1, .15), election_data = electdata)
}
