% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullmodel.R
\name{margingaussianmodelprop}
\alias{margingaussianmodelprop}
\title{Title}
\usage{
margingaussianmodelprop(
  poll_data,
  stateloc,
  marginloc,
  npolls = NULL,
  invgamma = F,
  v0 = NULL,
  a0 = NULL,
  b0 = NULL,
  election_data,
  cutoffs = c(-0.2, -0.1, -0.025, 0.025, 0.1, 0.2),
  groupnames = c("Strong Red", "Red", "Lean Red", "Competitive", "Lean Blue", "Blue",
    "Strong Blue")
)
}
\arguments{
\item{poll_data}{}

\item{stateloc}{}

\item{marginloc}{}

\item{invgamma}{}

\item{v0}{}

\item{a0}{}

\item{b0}{}

\item{election_data}{}

\item{cutoffs}{}

\item{groupnames}{}
}
\value{

}
\description{
Title
}
