% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.r
\name{propnorm}
\alias{propnorm}
\title{Proportional Normalization of data frame}
\usage{
propnorm(data)
}
\arguments{
\item{data}{}
}
\value{
A data frame that is proportionally normalized
}
\description{
Takes a data frame and normalizes it such that the rowSum is one
}
