% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.r
\name{propnormdfreplace}
\alias{propnormdfreplace}
\title{Title}
\usage{
propnormdfreplace(df, columns)
}
\arguments{
\item{df}{data frame to be}

\item{columns}{a vector of columns to be normalized}
}
\value{
a new data frame with both original and the normalized data with column names ending in normalized
}
\description{
Title
}
\examples{
df1 = data.frame(a = c(0.4, 0.3, .35, .45), b = c(.45, .40, .40, .3))
propnormdfreplace(df1)

df2  = data.frame(a = c(0.3, .34, .38), b = c(0.2, 0.21, 0.25), c = c(.45, .42, .41))
propnormdfreplace(df2)
}
