% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorcalculation.R
\name{rmse}
\alias{rmse}
\title{Root Mean Square Error}
\usage{
rmse(predicted, actual)
}
\arguments{
\item{predicted}{the predictions including the first column with a character vector with idenifiers to check that the rows of predicted and actual match}

\item{actual}{the real result including the first column with a character vector}
}
\value{
the Root Mean Square Error
}
\description{
This function calculates the root mean square error and checks that the idenifiers match. Note that if n=1, then the root mean square error is equivalent to the average error.
}
\examples{
predict1 = data.frame(state = c("TX", "GA", "OR"), dem = c(.45, .49, .63), rep = c(.55, .51, .37))
actual1 = data.frame(state = c("TX", "GA", "OR"), dem = c(.42, .48, .62), rep = c(.58, .52, .38))
rmse(predict1, actual1)
predict2 = data.frame(state = c("OK", "MS", "WY"), dem = c(.42, .44, .3), rep = c(.56, .55, .7), other = c(0.02, 0.01, 0))
actual2 = data.frame(state = c("OK", "MS", "WY"), dem = c(.41, .38, .3), rep = c(.55, .58, .69), other = c(0.04, 0.04, 0.01))
rmse(predict2, actual2)
}
