% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conjugatepriors.R
\name{unigausscp}
\alias{unigausscp}
\title{Title}
\usage{
unigausscp(data, priormean, priorvar, datavar = NULL, n = NULL,
  invgamma = F, V0 = NULL, a0 = NULL, b0 = NULL)
}
\arguments{
\item{data}{a vector  of the data or a single point}

\item{priormean}{mean of prior distribution}

\item{priorvar}{variance of prior distribution}

\item{datavar}{datavariance,  if null var is called to estimate it}

\item{n}{sample size}

\item{invgamma}{a logical indicator of whether or not a inverse gamma prior for sigma is (T) or isn't included}

\item{V0}{prior V0 where the likelihood is N(priormean)}

\item{a0}{prior a if gamma is T}

\item{b0}{prior b if gamma is T}
}
\description{
Title
}
