% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conjugatepriors.R
\name{unigausscpiterative}
\alias{unigausscpiterative}
\title{Iterative Gaussian Conjugate Prior
This function iteratives over a set of data.  The posterior becomes the new prior for the next data point}
\usage{
unigausscpiterative(data, priormean, priorvar, datavar = NULL, n,
  invgamma = F, V0 = NULL, a0 = NULL, b0 = NULL)
}
\arguments{
\item{priormean}{- intial prior mean}

\item{priorvar}{initial prior variance}

\item{datavar}{a vector of the variance for each data point}

\item{n}{a vector of sample sizes}

\item{invgamma}{a logical indicator of whether or not a inverse gamma prior for sigma is (T) or isn't included}

\item{V0}{initial v0}

\item{a0}{initial a0}

\item{b0}{inital b0}

\item{data-}{data for analysis}
}
\description{
Iterative Gaussian Conjugate Prior
This function iteratives over a set of data.  The posterior becomes the new prior for the next data point
}
