% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{fitbetavecMOM}
\alias{fitbetavecMOM}
\title{Beta Method of Moments
Calculates the method of monements estimators of a beta distribution fitted to the data}
\usage{
fitbetavecMOM(data)
}
\arguments{
\item{data}{- a vector of data}
}
\value{
The fitted a,b of a beta(a,b) distribtion
}
\description{
Beta Method of Moments
Calculates the method of monements estimators of a beta distribution fitted to the data
}
\examples{
set.seed(2)
data1 = rbeta(100, 1, 1)
fitbetavecMOM(data1)
data2 = rbeta(100, 43, 27)
fitbetavecMOM(data2)
}
