% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{fitnormalvecMLE}
\alias{fitnormalvecMLE}
\title{MLE estimates of a Normal Distribution
Fits a normal}
\usage{
fitnormalvecMLE(data)
}
\arguments{
\item{data}{- a vector of data}
}
\value{
a list with the elements mean with the mean of the distribution and
}
\description{
MLE estimates of a Normal Distribution
Fits a normal
}
