% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorgrouping.R
\name{getpriorassign}
\alias{getpriorassign}
\title{Title}
\usage{
getpriorassign(election_data, cutoffs = c(-0.2, -0.1, -0.025, 0.025, 0.1,
  0.2), groupnames = c("Strong Red", "Red", "Lean Red", "Competitive",
  "Lean Blue", "Blue", "Strong Blue"), weights = NULL)
}
\arguments{
\item{election_data}{the election data in df form. The first column must contain the name of that state}

\item{cutoffs}{the cutoffs used to split the data into the categories}

\item{weights}{optional weights for a weighted average}

\item{a}{character vector of names 
This function splits the data into categories based on the average of the margin, democratic vote, or difference}
}
\description{
Title
}
