% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullmodel.R
\name{margingaussianmodelprop}
\alias{margingaussianmodelprop}
\title{Title}
\usage{
margingaussianmodelprop(
  poll_data,
  stateloc,
  marginloc,
  nloc,
  proploc = NULL,
  npolls = NULL,
  dateloc = NULL,
  normalize = F,
  invgamma = F,
  v0 = NULL,
  a0 = NULL,
  b0 = NULL,
  election_data,
  cutoffs = c(-0.2, -0.1, -0.025, 0.025, 0.1, 0.2),
  groupnames = c("Strong Red", "Red", "Lean Red", "Competitive", "Lean Blue", "Blue",
    "Strong Blue")
)
}
\arguments{
\item{poll_data}{}

\item{stateloc}{}

\item{marginloc}{}

\item{npolls}{the number of polls to include, default is all polls}

\item{dateloc}{the location of the date used to determine the last npolls}

\item{normalize}{whether or not to normalize poll date before margin is calculated}

\item{invgamma}{}

\item{v0}{}

\item{a0}{}

\item{b0}{}

\item{election_data}{}

\item{cutoffs}{}

\item{groupnames}{}
}
\value{

}
\description{
Title
}
\examples{
require(politicaldata)
elect2008  = subset(pres_results , year == 2008)
elect2008$margin = elect2008$dem - elect2008$rep
elect2012 = subset(pres_results , year == 2012)
elect2012$margin = elect2012$dem - elect2012$rep
electdata = data.frame("state" = elect2008$state, "2008" =  elect2008$margin, "2012" = elect2012$margin)
data(polls2016)
polls = polls2016[ complete.cases(polls2016[, c(2, 3, 12)]), ]
stateloc  = which(colnames(polls) == "State")
polls2016$margin = polls2016$Trump - polls2016$Clinton
marginloc = which(colnames(polls2016) == "margin")
margingaussianmodelprop(polls2016, stateloc = stateloc, marginloc = marginloc, election_data = electdata)
}
