% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullmodel.R
\name{pollrollingaverage}
\alias{pollrollingaverage}
\title{Title}
\usage{
pollrollingaverage(poll_data, stateloc, proploc, candidateloc, npolls = 10)
}
\arguments{
\item{poll_data}{}

\item{stateloc}{}

\item{proploc}{}

\item{candidateloc}{}

\item{npolls}{}
}
\value{

}
\description{
Title
}
