% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.r
\name{propnorm}
\alias{propnorm}
\title{Proportional Normalization of data frame}
\usage{
propnorm(data)
}
\arguments{
\item{data}{the numeric data to be normailized}
}
\value{
A data frame that is proportionally normalized
}
\description{
Takes a data frame and normalizes it such that the rowSum is one
}
\examples{
df1 = data.frame(a = c(0.4, 0.3, .35, .45), b = c(.45, .40, .40, .3))
propnorm(df1)

df2  = data.frame(a = c(0.3, .34, .38), b = c(0.2, 0.21, 0.25), c = c(.45, .42, .41))
propnorm(df2)
}
