% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.r
\name{propnormdf}
\alias{propnormdf}
\title{Title}
\usage{
propnormdf(df, columns)
}
\arguments{
\item{df}{data frame to be}

\item{columns}{a vector of columns to be normalized}
}
\value{
a new data frame with the normalized data added with column names ending in normalized
}
\description{
Title
}
